backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
PyMySQL==1.1.1
python-dotenv==1.0.1
httpx==0.27.0
openai==1.40.0
a2wsgi==1.10.10
python-multipart==0.0.6
